Revert point class suggestions by Rene to define inline functions.
